adds marker creating and saving capability

+++ db/migrate/20150305213302_create_markers.rb
@@ -1,8 +1,8 @@
 class CreateMarkers < ActiveRecord::Migration
   def change
     create_table :markers do |t|
-      t.integer    'lat',  null: false
-      t.integer    'long', null: false
+      t.float    'lat',  null: false
+      t.float    'long', null: false
       t.string     'popup_content'
       t.references :map
       t.timestamps
